feat: change DATETIME64 default scale to 6, fix timestamp cast, and add comprehensive DateTime/DateTime64 UDF tests

+++ programs/local/PythonScalarUDF.cpp
@@ -692,7 +692,7 @@ void handleDatetime(
         case DB::TypeIndex::DateTime:
         {
             auto ts = value.attr("timestamp")();
-            column.insert(DB::Field(static_cast<UInt64>(ts.cast<uint64_t>())));
+            column.insert(DB::Field(static_cast<UInt64>(ts.cast<double>())));
             break;
         }
         case DB::TypeIndex::DateTime64:
